requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
pydantic==2.12.5
python-dotenv==1.0.0
openai>=1.40.0
pyodbc==5.3.0
